Added current project persistence in browser local storage

+++ editor-ui/src/main/java/com/github/lonelylockley/archinsight/components/RepositorySelectorComponent.java
@@ -3,6 +3,7 @@ package com.github.lonelylockley.archinsight.components;
 import com.github.lonelylockley.archinsight.MicronautContext;
 import com.github.lonelylockley.archinsight.events.BaseListener;
 import com.github.lonelylockley.archinsight.events.Communication;
+import com.github.lonelylockley.archinsight.events.RepositoryCloseEvent;
 import com.github.lonelylockley.archinsight.events.RepositorySelectionEvent;
 import com.github.lonelylockley.archinsight.model.remote.repository.RepositoryInfo;
 import com.github.lonelylockley.archinsight.remote.RemoteSource;
@@ -21,6 +22,8 @@ import com.vaadin.flow.server.VaadinService;
 import java.time.ZoneOffset;
 import java.time.format.DateTimeFormatter;
 import java.time.format.FormatStyle;
+import java.util.List;
+import java.util.UUID;
 
 public class RepositorySelectorComponent extends VerticalLayout {
 
@@ -54,6 +57,13 @@ public class RepositorySelectorComponent extends VerticalLayout {
             // this event is sent before a listener is added, so it won't cause cycle
             Communication.getBus().post(new RepositorySelectionEvent(null, item));
         }
+        else
+        if (items.size() > 1) {
+            restoreSelectedRepository(items);
+        }
+        else {
+            manageRepositoryButton.setText("<Create Repository>");
+        }
         if (!readOnly) {
             manageRepositoryButton.addClickListener(event -> {
                 var dlg = new ManagementDialog();
@@ -64,18 +74,51 @@ public class RepositorySelectorComponent extends VerticalLayout {
             manageRepositoryButton.setEnabled(false);
         }
 
+        final var repositoryCloseListener = new BaseListener<RepositoryCloseEvent>() {
+            @Override
+            @Subscribe
+            public void receive(RepositoryCloseEvent e) {
+                if (eventWasProducedForCurrentUiId(e)) {
+                    manageRepositoryButton.setText("<Choose Repository>");
+                    RepositorySelectorComponent.this.selected = null;
+                    storeSelectedRepository(null);
+                }
+            }
+        };
+        Communication.getBus().register(repositoryCloseListener);
+        addDetachListener(e -> { Communication.getBus().unregister(repositoryCloseListener); });
+
         final var repositorySelectionListener = new BaseListener<RepositorySelectionEvent>() {
             @Override
             @Subscribe
             public void receive(RepositorySelectionEvent e) {
-                manageRepositoryButton.setText(String.format("[ %s ]", e.getNewValue().getName()));
-                RepositorySelectorComponent.this.selected = e.getNewValue();
+                if (eventWasProducedForCurrentUiId(e)) {
+                    manageRepositoryButton.setText(String.format("[ %s ]", e.getNewValue().getName()));
+                    RepositorySelectorComponent.this.selected = e.getNewValue();
+                    storeSelectedRepository(e.getNewValue().getId());
+                }
             }
         };
         Communication.getBus().register(repositorySelectionListener);
         addDetachListener(e -> { Communication.getBus().unregister(repositorySelectionListener); });
 
         return manageRepositoryButton;
+    }
+
+    private void storeSelectedRepository(UUID repositoryId) {
+       getElement().executeJs("localStorage.setItem($0, $1)", "org.archinsight.editor.project", repositoryId == null ? null : repositoryId.toString());
+    }
+
+    private void restoreSelectedRepository(List<RepositoryInfo> items) {
+        getElement().executeJs("return localStorage.getItem($0)", "org.archinsight.editor.project").then(String.class, repositoryId -> {
+            if (repositoryId != null) {
+                var uuid = UUID.fromString(repositoryId);
+                items.stream().filter(repo -> repo.getId().equals(uuid)).forEach(repo -> {
+                    Communication.getBus().post(new RepositorySelectionEvent(RepositorySelectorComponent.this.selected, repo));
+                    RepositorySelectorComponent.this.selected = repo;
+                });
+            }
+        });
     }
 
     private class ManagementDialog extends Dialog {
@@ -97,6 +140,7 @@ public class RepositorySelectorComponent extends VerticalLayout {
             var selectButton = new Button("Select", e -> {
                 var selection = table.getSelectedItems();
                 if (selection.size() > 0) {
+                    Communication.getBus().post(new RepositoryCloseEvent());
                     Communication.getBus().post(new RepositorySelectionEvent(RepositorySelectorComponent.this.selected, selection.iterator().next()));
                     close();
                 }
@@ -123,10 +167,16 @@ public class RepositorySelectorComponent extends VerticalLayout {
             deleteButton.setMinWidth("80px");
             deleteButton.setEnabled(false);
             deleteButton.addClickListener(e -> {
-                table.getSelectedItems().forEach(repo -> {
-                    remoteSource.repository.removeRepository(repo.getId());
-                    table.getListDataView().removeItem(repo);
-                });
+                var items = table.getSelectedItems();
+                if (items.size() > 0) {
+                    items.forEach(repo -> {
+                        if (repo.equals(RepositorySelectorComponent.this.selected)) {
+                            Communication.getBus().post(new RepositoryCloseEvent());
+                        }
+                        remoteSource.repository.removeRepository(repo.getId());
+                        table.getListDataView().removeItem(repo);
+                    });
+                }
             });
             controls.add(input);
             controls.add(createButton);
